backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
beautifulsoup4
requests
python-dotenv
pytest
